fixup! Refactor: prevent start/stop dev server from entering debugger
When running start-dev-server or stop-dev-server for the first time upon a fresh
(asdf:load-system :project-isidore) inside a new REPL will yield,

There is no applicable method for the generic function
/#<STANDARD-GENERIC-FUNCTION HUNCHENTOOT::ACCEPTOR-LISTEN-SOCKET (1)> when called
with arguments (NIL). [Condition of type SB_PCL::NO-APPLICABLE-METHOD-ERROR]

This is due to the HUNCHENTOOT:STARTED-P having no methods applicable to
when *app-dev* is set to nil in its initial declaration. See
[[file:~/quicklisp/dists/quicklisp/software/hunchentoot-v1.3.0/acceptor.lisp::defgeneric
started-p (acceptor][<no description>]]

To account for this special case we use (equalp *app-dev* nil) which returns t
to signify this. Then the standard control structure UNLESS takes care of the rest.