Refactor: Move heroku.lisp into compile.lisp. New /src folder
Previously heroku.lisp sole purpose was to redefine the initialize-application
symbol first defined in bin/compile.lisp. compile.lisp has been changed so the
redefinition is no longer needed.

Historically the reason for the structure was my unfamiliarity with the
buildpack itself.

---

The Heroku build pack that works with SBCL implementation can be found here:
[[https://gitlab.com/bendersteed/heroku-buildpack-common-lisp][bendersteed / heroku-buildpack-common-lisp · GitLab]]  ([[https://bendersteed.gitlab.io/post/deploying-common-lisp-to-heroku/][see Author's journal of
the buildpack fork]]). In case the blog goes down, the following is reiterated.
Relevant caveats include that in order for the buildpack to work, you have to
define the initialize-application symbol in CL-USER package. This is not really
good style, I would pretty much prefer to be able to have the function in my
package (aka run in the REPL, =in-package :project-isidore=). For now I
workaround this caveat by adding a heroku.lisp file to my system.