Build: Remove support for Clozure Common Lisp
In preparation to simplify and understand my build process, I am removing
support for CCL (& Roswell in the upcoming commits). If I want to add
portability in the future it will be through less #+sbcl or #+ccl conditionals
but rather through the ever expanding mission creep of ASDF.

CI is another story. Core app logic should be implementation agnostic and here
Roswell does a great service to testing. As per my current use case, I don't
foresee deploying on more than SBCL likely in the near future. Therefore as
barebones as my application already is, I desire to perfect it more by carving
away the fat.

Taking out the complexity of Roswell so that the build is as simple as needs be
should save me some time in the future. Fingers crossed.