Build: Combine initialize-application and start-dev-server
INITIALIZE-APPLICATION used to be for production purposes and START-DEV-SERVER
for development purposes. START-DEV-SERVER has been combined into
INITIALIZE-APPLICATION with a keyword argument PRODUCTIONP to account for
differences which are two at this time of writing

1. Generated static assets must be output to different pathnames
2. In a production environment, the initialization of the application must
retrieve the Heroku PostgreSQL database URL

Consequently, this refactor means a reduction in the amount of exported symbols
and functions.