Refactor: Move from XHTML to HTML5 document formatting
I do think I enjoy history more than programming. CL-WHO uses XHTML by default,
and so we switch to HTML5 for access to some feature tags like <video> etc. Who
doesn't want to adhere to web standards?

From /u/remy_porter @ https://old.reddit.com/r/webdev/comments/2qxxu1/html5_vs_xhtml/

In the beginning was SGML, the Standard Generalized Markup Language. SGML was
powerful, and people said, "Hey, can we use this for document layout?" And thus
was born HTML. A different group of people said, "Hey, can we use this for data
storage? With validating schemas?" And thus was born XML.

XML and HTML were both children of SGML, but they existed to fulfill very
different purposes. HTML was designed to be sloppy and forgiving. At no point
should a poorly written HTML document fail to render something, even if what it
renders is inaccurate and awful. It was a presentational tool, not a data tool.

XML on the other hand, as a data tool, was meant to be much more strict. A
poorly formed XML document would simply fail to load. An XML document that
didn't map to its schema would chuck out all sorts of ugly errors.

In the early 2000s, XML was all the rage. The aftermath of the browser wars left
us all struggling to write sites that actually worked, and so some people said,
"What if we made a subset of XML that looks like HTML, but has to be well formed
and validate against a schema?" And thus XHTML was born.

Let's recap that geneology real quick. SGML begat XML and HTML. XML begat XHTML.
XHTML was meant to look like HTML, but it is not, itself, descended from HTML.
The parsers were much simpler to build, because the language was much more
strict, but nobody was going to rebuild their sites to take advantage of it, so
everybody still needed the less strict (and more complex) HTML parser, so
really, XHTML didn't change anything.

Fast forward to HTML5. HTML5 is not specified as a descendant of SGML anymore.
It mostly supports an HTML-style or XHTML-style syntax, and doesn't particularly
care which is which, however it does have a subspecification for XHTML5, which
is meant to be HTML5 serialized in XML (for use in web services).

Now, to get to your question:

    I was wondering which language would be better these days to program in?

Neither, because they're both markup languages, not programming languages. But
for markup purposes, you should use HTML5. Whether you use XHTML-syntax or not
is up to you, but by general convention you do not.