CI: Enable Continuous Delivery for lispapp executable image dump
Use Github Actions to build our application binary on GNU/Linux and MacOS.
Microsoft Windows is not supported currently. See the Work-in-Progress
'windows-binary' branch. The generated artifacts can then be downloaded from the
Github actions.

lispapp.ros is used as it proved too much work to modify compile.lisp to
accommodate Roswell's idiosyncrasies. It is also CL implementation and Operating
System agnostic.

The build process with Ubuntu on Heroku is still the "blessed" implementation.