CI: Refactor CD.yml to not use matrix, OS specific jobs instead.
This is in preparation for adding the windows executable build. It relies on the
Github Action msys2/setup-msys2@v2 to mimic a UNIX-like environment to build software.

This action will error out with the message of wrong runner environment when
combined into the current structure. Windows build will require its own job:
field.