CI: CD Add building executable on Microsoft Windows
:)

N.B.:

A) Common Lisp Library CI-Utils is outdated.

As of this commit date, https://github.com/neil-lindquist/CI-Utils has not been
updated to reflect the change in the upstream Roswell repo in this particular
commit:

https://github.com/roswell/roswell/commit/fd3b28216424614e67be8ec3d946b9bd2b4bde54#diff-cc067a0ad4d3c0ba5c053fc05c9920e44d7c230edf5585acabc50c8e163e36dc

where 'pacman' is used to retrieve the Windows version of Roswell. That commit
could use more context in explaining that CI with Roswell now requires the
'msys2/setup-msys2@v2' Github Action to install msys2. This additional
dependency on msys2 means Mr. Lindquist's very helpful example configs need
updating to reflect this fact. I'll open a PR should I have time in the future.

B) Default behavior of Roswell build subcommand

from  https://roswell.github.io/Building-images-and-executables.html,

executable NAME [-o OUTPUT]

When a script is dumped with executable, the dumped image becomes an
self-contained executable binary which implies --restart main.

If OUTPUT is given, the resulting binary is written to this file. Otherwise, the
output filename is deduced from NAME and is written in the same directory. On
Windows and if SCRIPT has .ros extension, the filename will be SCRIPT.exe. On
the other systems, the result will be SCRIPT (without extension).

This feature is supported on SBCL, CCL, CMUCL, CLISP, ECL.

The flag DOES NOT work. Or I could not get it to work on CD.yml. Not a big
issue, we can rely on default behavior. So note that on Windows it will be
'lispapp.exe' instead of 'lispapp'.

C) The following commands are unable to be executed:

    - name: Run Lisp
      continue-on-error: true
      shell: bash
      run: |
        ros -e '(format t "~a:~a on ~a~%...~%~%" (lisp-implementation-type) (lisp-implementation-version) (machine-type))'
        ros -e '(format t " fixnum bits:~a~%" (integer-length most-positive-fixnum))'
        ros -e "(ql:quickload 'trivial-features)" -e '(format t "features = ~s~%" *features*)'

    - name: Update Quicklisp distribution if cached
      shell: bash
      run: ros -e "(ql:update-all-dists :prompt nil)"

but 'ros build bin/lispapp.ros' works. The error message that the former
commands can't be executed because the 'ros' command cannot be found is not
helpful. Too much effort right now to track down the error.