CI: Rename build artifacts. Remove obsolete env var MATRIX.LISP.
from:
project-isidore-Linux-x64-v1.0.0

to:
pi-build-Linux

Now the build artifact consists of more than one file, but rather a directory tree.
Files like LICENSE and assets/ are included

Note that https://github.com/softprops/action-gh-release only allows releasing
file(s) and not directories.

So we must .zip the build directory to produce a single file. Therefore the
OS, architecture, and version information ought to be the name of the .zip file.
Naming the build artifact in such a way is redundant and confusing.

Upon release, the contents of pi-build-Linux/* will be zipped up and renamed.

CD.yml is no longer using a matrix strategy, so ${{ matrix.lisp }} is irrelevant.