Refactor: Split single .asd file into two subsystems.
Apparently for package inferred systems, the .asd file needs to be placed inside
the project directory and not the parent directory.

See: comments and later additions on this article: https://davazp.net/2014/11/26/modern-library-with-asdf-and-package-inferred-system.html

This change better reflects the fact that project-isidore-tests is really a
distinct but not separate system. Another benefit is the shorter and more accurate package names.