CI: Remove Roswell from Windows build process.
I needed to test the bknr.datastore library with CCL on the x86-64 Windows
platform as SBCL with x86-64 on Windows was failing to load the library.

As per the issue here: https://github.com/roswell/roswell/issues/491 , Roswell
is unable to install 'ccl-bin' on Windows due to some missing win32 headers. The
Roswell project author has confirmed the bug and as of the current date, no
workaround has been found.

I was able to install CCL on Windows without Roswell fairly simply. When it
turned out that CCL was unable to support the bknr.datastore library also, it
was determined some modifications needed to be made to the library itself.
Having made the modifications needed for the library to run under SBCL x86-64
Windows, I had already laid the groundwork for building a Windows binary on
Github Actions without Roswell for CCL. Instead of putting Roswell back in, I
decided that since "in Rome, do as the Romans do", I would script the CD
pipeline to use powershell for Windows and SBCL.

CD build process on Windows is now the same speed as Linux and MacOS, thanks to
cutting out the msys2 portion of the pipeline.

I support the goal of the Roswell authors, and gave their software an honest
try. As it stands, it fails in being a lisp implementation manager for any of
the less common environments, namely Windows. My uses of Roswell is also very
trivial, not extending far beyond the given examples.

See commit 1d181e4 also for an example of brittleness involving any Roswell
scripting on MacOS.

See commit bf210b8 for another example of Roswell scripting failing to work with
ASDF package inferred systems.

I will continue to use Roswell for CI as it works well there. With enough time
I am sure Roswell will live up to its stated purpose: a lisp implementation
manager that just works.

For powershell command info, please see:

https://social.technet.microsoft.com/Forums/en-US/23fccc30-5f84-4a84-8160-c6e95102b11c/powershell-invokewebrequest-sourceforge-urlsredirectiondynamic-content?forum=winserverpowershell

https://stackoverflow.com/questions/8839978/install-msi-with-msiexec-into-specific-directory

https://powershellexplained.com/2016-10-21-powershell-installing-msi-files/

https://www.advancedinstaller.com/forums/viewtopic.php?t=45223