Refactor: Generate system dependencies.
The point of using package inferred systems is to alleviate the need to manually
specify loading order in a package.lisp file. However said file is useful to
view at once all the external dependencies of a system. With a package inferred
system it will have to be generated.

View all external dependencies by calling LIST-PROJECT-ISIDORE-DEPENDENCIES.