Feat: Generate links to currently displayed chapters in the Bible.
This set of links needs to be generated dynamically. Therefore unlike the book
links, the chapter links take the URL as input and the function
MAKE-BIBLE-CHAPTER-URL-LIST in turn generates the right links for the URL.

The associative list for the chapters needed also to be extracted from the data
structure. See the changes in migration.lisp for that.