Refactor: Fix missing in-package declaration. Prefer define-package.
Changes originally introduced in 64b555b and bf210b8.

uiop:define-package allows use of the :mix and :reexport clauses. The difference
between :use and :mix is that :use says "I want all the symbols of the listed
packages, conflicts are to be resolved manually", while :mix says "I want all
the symbols of the listed packages, conflicts are to be resolved in favour of
the earlier listed symbol".