Fix: Bad import of commentary. Wrong header in source file.
When I was using regex search and replace to prepend titles with *, II Peter got
mistakenly renamed as II Peter. This however only happened to the Book title,
and not the Chapter and Verses. This omission manifested when I notice II Peter
had no commentaries displayed on the web interface.