Refactor: Move association lists to data.lisp. Use DEFCONSTANT.
I have tried putting these in their own class and slot in BKNR.DATASTORE,
however this results in a 100% percent increase in performance cost.

When using defconstant, there are two common pitfalls.

1. Know that `eql' is using to determine equality (as opposed to Common
Lisp's other equality functions like =, eq, equal, or equalp)

From:
https://stackoverflow.com/questions/34800988/sbcl-asdfload-system-fails-when-a-string-constant-is-defined

Why does the string constant fails?

The specification for defconstant tells us that:

    However, the consequences are undefined if an attempt is made to assign a value to the symbol using another operator, or to assign it to a different value using a subsequent defconstant.

The important word here is different: according to which comparison?

    The consequences are undefined if there are any bindings of the variable named by name at the time defconstant is executed or if the value is not eql to the value of initial-value.

The comparison is done by eql.

SBCL compiles your file and then loads the result (the xxx-TMP.fasl file), and for that particular implementation, the defconstant form is thus evaluated twice in the same environment. A compiler needs not actually evaluate the form during compilation (it could declare it internally somehow so that further uses of the constants can be inlined) but this is a valid compilation strategy.

Here, since the compilation environment and the load environment are the same, and because the two occurences of the strings are not identical (not eq), an error is signaled. If you happened to load the FASL file with another instance of the same version of the SBCL interpreter, it would not give you this error.
What can you do?

    Don't reinvent the wheel, use alexandria:define-constant, which allows to specify under which test function the value is constant:

    (alexandria:define-constant b "B" :test #'string=)

    Thus, when it is evaluated multiple times, the new "B" value is compared against the existing value using string=, and since they are equal, no other action is performed.

    Use defvar and do not worry about it anymore until you deploy your code (often constants need to change during development).

From: https://stackoverflow.com/questions/69778914/unable-to-compile-file-when-a-defconstant-uses-a-function-from-the-same-file

2. Unable to compile file when a defconstant uses a function from the same file

(defun f (x)
  (* x 2))

(defconstant +n+ (f 1))

From DEFCONSTANT:

    [...] users must ensure that the initial-value can be evaluated at compile time (regardless of whether or not references to name appear in the file) and that it always evaluates to the same value.

But COMPILE-FILE and 3.2.3 File Compilation only require the DEFUN form to establish that F exists, not that is may be evaluated during compilation.

You need to LOAD the resulting file for the definition to be available in your environment (e.g. Slime compiles a file and loads the result). Likewise, when you evaluate forms one at a time in the REPL, the definition becomes available right away, as-if a small compile-load cycle was done.

You have two main options:

    wrap the defun in an eval-when form

      (eval-when (:compile-toplevel :load-toplevel :execute)
        (defun f (x) (* x 2)))

    put the two forms in distinct files, and have the second one depends in the system (in the defsystem sense) on the first one:

      (defsystem "foo"
        :components ((:file "define-f")
                     (:file "define-constant"))
        :serial t)