Feat: Create regex based HTML links to Haydock cross references.
Possibly implement a visit on hover with client side javascript in the future.
Regex used is still not perfect and will have to be tweaked in the future. It
errs on the side of missing valid links over creating invalid ones.

Roman numeral code taken from
https://stackoverflow.com/questions/40515523/lisp-check-roman-numeral-converter-for-valid-roman-numeral
where I added the roman-numeral-p function.

A few tests were written. MAKE.LISP was changed to accommodate the new tests.