Build: Avoid Windows binary dependency.
When running the Windows binary you may run into an error similar to,

("libcrypto-1_1-x64.dll" "libeay32.dll") [Condition of type CFFI:LOAD-FOREIGN-LIBRARY-ERROR]

Project Isidore uses the
[[https://common-lisp.net/project/cl-plus-ssl/][cl+ssl]] library which in turn
requires OpenSSL as a dependency if the web server is to send HTTPS requests.
Hit =0= to continue twice and the application will proceed as normal. Firewall
network exceptions need not be granted if the application is to be used locally.

Now I have done as the Hunchentoot documentation says and pushed the appopriate
symbol to the *features* list.